Have the plugin precisely mimic the CLI tool
This is a terrible hack, but it appears to be the only way to have the plugin
organize imports as well. The google-java-format CLI tool unconditionally
invokes ImportOrderer, which is package-private. This invocation is performed
by FormatFileCallable, but instances of that class can only be created by
passing arguments of the package-private type CommandLineOptions. So for the
time being we'll just close our eyes.